+++ tests/Type/GeneralTypeTest.php
@@ -25,6 +25,7 @@ class GeneralTypeTest extends TypeInferenceTestCase
         yield from $this->gatherAssertTypes(__DIR__.'/data/environment-helper.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/view.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/bug-1346.php');
+        yield from $this->gatherAssertTypes(__DIR__.'/data/request-header.php');
     }
 
     /**